perf: 启用 BoringSSL AES-NI 汇编加速——MinGW 兼容性 patch
MinGW + Windows 下 BoringSSL 的 fiat_p256_adx .S 文件不支持 PE/COFF 格式，
导致 OPENSSL_NO_ASM 被设置为 ON，禁用了所有汇编优化（包括 AES-NI）。

修改方案：
1. CMakeLists.txt: 使用 FetchContent_Populate + patch 移除 fiat_p256_adx 文件
2. p256_64.h: 添加 !defined(OPENSSL_WINDOWS) 条件避免调用未实现的函数

性能提升：
- AES-256-GCM: 205 Mi/s → 13.7 Gi/s（67 倍提升）
- X25519 密钥交换: 81.5 us → 21.4 us（3.8 倍提升）
- 更新性能报告文档反映新数据

+++ benchmarks/CodecBench.cpp
@@ -114,7 +114,7 @@ static void BM_HttpParseProxyRequest_PostBody(benchmark::State &state)
         }
         benchmark::DoNotOptimize(req);
     }
-    state.SetBytesProcessed(int64_t(state.iterations()) * int64_t(payload.size()));
+    state.SetBytesProcessed(state.iterations() * static_cast<int64_t>(payload.size()));
 }
 
 static void BM_HttpExtractRelativePath(benchmark::State &state)
@@ -149,7 +149,7 @@ static void BM_TrojanDecodeCredential(benchmark::State &state)
         }
         benchmark::DoNotOptimize(credential);
     }
-    state.SetBytesProcessed(static_cast<std::int64_t>(state.iterations()) * static_cast<std::int64_t>(buffer.size()));
+    state.SetBytesProcessed(state.iterations() * static_cast<std::int64_t>(buffer.size()));
 }
 
 static void BM_TrojanDecodeCredential_Invalid(benchmark::State &state)
@@ -163,7 +163,7 @@ static void BM_TrojanDecodeCredential_Invalid(benchmark::State &state)
         benchmark::DoNotOptimize(ec);
         benchmark::DoNotOptimize(credential);
     }
-    state.SetBytesProcessed(static_cast<std::int64_t>(state.iterations()) * static_cast<std::int64_t>(buffer.size()));
+    state.SetBytesProcessed(state.iterations() * static_cast<std::int64_t>(buffer.size()));
 }
 
 // ============================================================
@@ -188,7 +188,7 @@ static void BM_DnsPackMessage(benchmark::State &state)
         auto wire = msg.pack();
         benchmark::DoNotOptimize(wire);
     }
-    state.SetBytesProcessed(static_cast<std::int64_t>(state.iterations()) * 64);
+    state.SetBytesProcessed(state.iterations() * 64);
 }
 
 static void BM_DnsUnpackMessage(benchmark::State &state)
@@ -200,7 +200,7 @@ static void BM_DnsUnpackMessage(benchmark::State &state)
         auto opt = resolve::message::unpack(std::span<const std::uint8_t>(wire.data(), wire.size()));
         benchmark::DoNotOptimize(opt);
     }
-    state.SetBytesProcessed(static_cast<std::int64_t>(state.iterations()) * static_cast<std::int64_t>(wire.size()));
+    state.SetBytesProcessed(state.iterations() * static_cast<std::int64_t>(wire.size()));
 }
 
 static void BM_DnsExtractIps(benchmark::State &state)
@@ -255,7 +255,7 @@ static void BM_Sha224Short(benchmark::State &state)
         auto hash = crypto::sha224("abc");
         benchmark::DoNotOptimize(hash);
     }
-    state.SetBytesProcessed(static_cast<std::int64_t>(state.iterations()) * 3);
+    state.SetBytesProcessed(state.iterations() * 3);
 }
 
 static void BM_Sha224Long(benchmark::State &state)
